first refactor for unittests. made them more flexible and appendable. fixes #253
