+++ tests/start.py
@@ -1,0 +1,13 @@
+import time
+import unittest
+
+from unittests import TestBasicChat as test_a
+from unittests import TestAPI as test_b
+from unittests import TestCommands as test_c
+
+# TODO: import new unittests in alphabetical order
+
+
+if __name__ == '__main__':
+    time.sleep(3)
+    unittest.main()
