+++ tests/test_objs/__init__.py
@@ -1,0 +1,1 @@
+from .sockets import SocketIOConnection
