+++ tests/unittests/__init__.py
@@ -1,0 +1,3 @@
+from .socket_chat import TestBasicChat
+from .api import TestAPI
+from .socket_commands import TestCommands
